requirements.txt
pyrogram
tgcrypto
google-generativeai
Pillow
nest_asyncio
flask
